adds profile component
profile component will contain user settings (right now only category change)
